+++ R/external-generic.R
@@ -40,6 +40,20 @@ new_external_generic <- function(package, name, dispatch_args, version = NULL) {
   out
 }
 
+as_external_generic <- function(x) {
+  if (is_S7_generic(x)) {
+    pkg <- package_name(x)
+    new_external_generic(pkg, x@name, x@dispatch_args)
+  } else if (is_external_generic(x)) {
+    x
+  } else if (is_S3_generic(x)) {
+    pkg <- package_name(x)
+    new_external_generic(pkg, x$name, "__S3__")
+  } else if (is_S4_generic(x)) {
+    new_external_generic(x@package, as.vector(x@generic), x@signature)
+  }
+}
+
 #' @export
 print.S7_external_generic <- function(x, ...) {
   cat(
@@ -67,13 +81,14 @@ is_external_generic <- function(x) {
 #'
 #' @importFrom utils getFromNamespace packageName
 #' @export
+#' @returns Nothing; called for its side-effects.
 #' @examples
 #' .onLoad <- function(...) {
 #'   S7::methods_register()
 #' }
 methods_register <- function() {
   package <- packageName(parent.frame())
-  tbl <- external_methods_get(package)
+  tbl <- S7_methods_table(package)
 
   for (x in tbl) {
     register <- registrar(x$generic, x$signature, x$method)
@@ -84,6 +99,8 @@ methods_register <- function() {
       setHook(packageEvent(x$generic$package, "onLoad"), register)
     }
   }
+
+  invisible()
 }
 
 registrar <- function(generic, signature, method) {
@@ -104,28 +121,34 @@ registrar <- function(generic, signature, method) {
   }
 }
 
-external_methods_get <- function(package) {
-  S3_methods_table(package)[[".S7_methods"]] %||% list()
-}
-
 external_methods_reset <- function(package) {
-  tbl <- S3_methods_table(package)
-  tbl[[".S7_methods"]] <- list()
+  S7_methods_table(package) <- list()
   invisible()
 }
 
 external_methods_add <- function(package, generic, signature, method) {
-  tbl <- S3_methods_table(package)
+  tbl <- S7_methods_table(package)
 
   methods <- append(
-    tbl[[".S7_methods"]] %||% list(),
+    tbl,
     list(list(generic = generic, signature = signature, method = method))
   )
 
-  tbl[[".S7_methods"]] <- methods
+  S7_methods_table(package) <- methods
   invisible()
 }
 
-S3_methods_table <- function(package) {
-  asNamespace(package)[[".__S3MethodsTable__."]]
+# Store external methods in an attribute of the S3 methods table since
+# this mutable object is present in all packages.
+
+S7_methods_table <- function(package) {
+  ns <- asNamespace(package)
+  tbl <- ns[[".__S3MethodsTable__."]]
+  attr(tbl, "S7methods")
+}
+`S7_methods_table<-` <- function(package, value) {
+  ns <- asNamespace(package)
+  tbl <- ns[[".__S3MethodsTable__."]]
+  attr(tbl, "S7methods") <- value
+  invisible()
 }