+++ R/property.R
@@ -12,18 +12,31 @@
 #' @param class Class that the property must be an instance of.
 #'   See [as_class()] for details.
 #' @param getter An optional function used to get the value. The function
-#'   should take `self`  as its sole argument and return the value. If the
-#'   property has a `class` the class of the value is validated.
+#'   should take `self` as its sole argument and return the value. If you
+#'   supply a `getter`, you are responsible for ensuring that it returns
+#'   an object of the correct `class`; it will not be validated automatically.
 #'
 #'   If a property has a getter but doesn't have a setter, it is read only.
 #' @param setter An optional function used to set the value. The function
 #'   should take `self` and `value` and return a modified object.
+#' @param validator A function taking a single argument, `value`, the value
+#'   to validate.
+#'
+#'   The job of a validator is to determine whether the property value is valid.
+#'   It should return `NULL` if the object is valid, or if it's not valid,
+#'   a single string describing the problem. The message should not include the
+#'   name of the property as this will be automatically appended to the
+#'   beginning of the message.
+#'
+#'   The validator will be called after the `class` has been verified, so
+#'   your code can assume that `value` has known type.
 #' @param default When an object is created and the property is not supplied,
 #'   what should it default to? If `NULL`, defaults to the "empty" instance
 #'   of `class`.
-#' @param name Property name, primarily used for error messages. Used
-#'   primrarily for testing as it is set automatically when using a list of
-#'   properties.
+#' @param name Property name, primarily used for error messages. Generally
+#'   don't need to set this here, as it's more convenient to supply as a
+#'   the element name when defining a list of properties. If both `name`
+#'   and a list-name are supplied, the list-name will be used.
 #' @returns An S7 property, i.e. a list with class `S7_property`.
 #' @export
 #' @examples
@@ -68,7 +81,12 @@
 #' hadley@firstName
 #' hadley@firstName <- "John"
 #' hadley@first_name
-new_property <- function(class = class_any, getter = NULL, setter = NULL, default = NULL, name = NULL) {
+new_property <- function(class = class_any,
+                         getter = NULL,
+                         setter = NULL,
+                         validator = NULL,
+                         default = NULL,
+                         name = NULL) {
   class <- as_class(class)
   if (!is.null(default) && !class_inherits(default, class)) {
     msg <- sprintf("`default` must be an instance of %s, not a %s", class_desc(class), obj_desc(default))
@@ -81,12 +99,16 @@ new_property <- function(class = class_any, getter = NULL, setter = NULL, defaul
   if (!is.null(setter)) {
     check_function(setter, alist(self = , value = ))
   }
+  if (!is.null(validator)) {
+    check_function(validator, alist(value = ))
+  }
 
   out <- list(
     name = name,
     class = class,
     getter = getter,
     setter = setter,
+    validator = validator,
     default = default
   )
   class(out) <- "S7_property"
@@ -189,19 +211,22 @@ prop_obj <- function(object, name) {
       stop(msg, call. = FALSE)
     }
 
-    if (isTRUE(check) && !class_inherits(value, prop$class)) {
-      stop(prop_error_type(object, name, prop$class, value), call. = FALSE)
-    }
-
     if (!is.null(prop$setter) && !identical(setter_property, name)) {
       setter_property <<- name
       on.exit(setter_property <<- NULL, add = TRUE)
       object <- prop$setter(object, value)
     } else {
+      if (isTRUE(check)) {
+        error <- prop_validate(prop, value, object)
+        if (!is.null(error)) {
+          stop(error, call. = FALSE)
+        }
+      }
+
       attr(object, name) <- value
     }
 
-    if (isTRUE(check)) {
+    if (isTRUE(check) && is.null(setter_property)) {
       validate(object, properties = FALSE)
     }
 
@@ -213,20 +238,28 @@ prop_error_unknown <- function(object, prop_name) {
   sprintf("Can't find property %s@%s", obj_desc(object), prop_name)
 }
 
-prop_error_type <- function(object, prop_name, expected, actual, show_type = TRUE) {
-  sprintf("%s@%s must be %s, not %s",
-    if (show_type) obj_desc(object) else "",
-    prop_name,
-    class_desc(expected),
-    obj_desc(actual)
-  )
+prop_validate <- function(prop, value, object = NULL) {
+  if (!class_inherits(value, prop$class)) {
+    sprintf("%s must be %s, not %s",
+      prop_label(object, prop$name),
+      class_desc(prop$class),
+      obj_desc(value)
+    )
+  } else if (!is.null(prop$validator)) {
+    val <- prop$validator(value)
+    if (!is.null(val)) {
+      paste0(prop_label(object, prop$name), " ", val)
+    } else {
+      NULL
+    }
+  } else {
+    NULL
+  }
 }
 
-#' @rdname prop
-#' @usage object@name
-#' @aliases @
-#' @rawNamespace if (getRversion() >= "4.3.0") S3method(base::`@`, S7_object) else export("@")
-`@.S7_object` <- prop
+prop_label <- function(object, name) {
+  sprintf("%s@%s", if (!is.null(object)) obj_desc(object) else "", name)
+}
 
 # Note: we need to explicitly refer to base with "base::`@`" in the
 # namespace directive to ensure the method is registered in the correct place.
@@ -234,16 +267,10 @@ prop_error_type <- function(object, prop_name, expected, actual, show_type = TRU
 # presence of a closure w/ the name of the generic (`@`) in the R7 namespace,
 # and incorrectly assumes that R7::`@` is the generic and registers the
 # method in the package namespace instead of base::.__S3MethodsTable__.
-
-`@` <- function(object, name) {
-  if (inherits(object, "S7_object")) {
-    name <- as.character(substitute(name))
-    prop(object, name)
-  } else {
-    name <- substitute(name)
-    do.call(base::`@`, list(object, name))
-  }
-}
+#' @usage object@name
+#' @rawNamespace if (getRversion() >= "4.3.0") S3method(base::`@`, S7_object)
+#' @name prop
+`@.S7_object` <- prop
 
 #' @rawNamespace S3method("@<-",S7_object)
 `@<-.S7_object` <- function(object, name, value) {
@@ -367,7 +394,7 @@ as_properties <- function(x) {
   }
 
   out <- Map(as_property, x, names2(x), seq_along(x))
-  names(out) <- names2(x)
+  names(out) <- vapply(out, function(x) x$name, FUN.VALUE = character(1))
 
   if (anyDuplicated(names(out))) {
     stop("`properties` names must be unique", call. = FALSE)
@@ -377,15 +404,23 @@ as_properties <- function(x) {
 }
 
 as_property <- function(x, name, i) {
-  if (name == "") {
-    msg <- sprintf("`property[[%i]]` is missing a name", i)
-    stop(msg, call. = FALSE)
-  }
 
   if (is_property(x)) {
-    x$name <- name
+    if (name == "") {
+      if (is.null(x$name)) {
+        msg <- sprintf("`properties[[%i]]` must have a name or be named.", i)
+        stop(msg, call. = FALSE)
+      }
+    } else {
+      x$name <- name
+    }
     x
   } else {
+    if (name == "") {
+      msg <- sprintf("`properties[[%i]]` must be named.", i)
+      stop(msg, call. = FALSE)
+    }
+
     class <- as_class(x, arg = paste0("property$", name))
     new_property(x, name = name)
   }